# services/tracking-go/internal/openapi/testdata/python-contract-at-cutover.yaml
...
openapi: 3.1.0
info:
  title: Tracking Service API
  description: 'HTTP API for the 3MRAI Tracking microservice (FastAPI + Aurora MySQL). Three
    distinct auth schemes, documented per route rather than globally: the user-scoped reads
    and creation take the gateway-injected x-user-id header (which carries the Cognito sub,
    not the internal usr_ id); the carrier status update takes its own external x-api-key;
    health and the E2E cleanup take neither.'
  version: 1.0.0
servers:
- url: http://localhost:3002
  description: Local (docker compose / Floci)
paths:
...
  /v1/trackings/by-user:
    delete:
      tags:
      - internal
      summary: '[Internal] Soft-delete every tracking belonging to a user'
      description: Soft-delete the user's trackings and, through the FK, their history.
      operationId: delete_trackings_by_user_v1_trackings_by_user_delete
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalDeleteByUserRequest'
      responses:
        '200':
          description: The user's trackings and their history are soft-deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalDeleteByUserResponse'
        '401':
          description: Missing or invalid internal API key
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/trackings:
    get:
      tags:
      - trackings
      summary: Read several of the caller's trackings by order id
      description: 'Return the caller''s trackings among `order_ids`, each with its history.


        Always `200`, even when nothing matches: "none of these are yours" is a

        complete answer to the question asked, not a failure. An empty list is the

        correct body, and it is deliberately indistinguishable from "none of these

        exist" — see the ownership rule in the module docstring.'
      operationId: get_trackings_v1_trackings_get
      parameters:
      - name: order_ids
        in: query
        required: true
        schema:
          type: string
          description: Comma-separated order ids, e.g. `ord_a,ord_b`. Ids the caller does
            not own are omitted from the response.
          title: Order Ids
        description: Comma-separated order ids, e.g. `ord_a,ord_b`. Ids the caller does not
          own are omitted from the response.
      - name: x-user-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-User-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingListResponse'
        '400':
          description: More than 100 order_ids
        '401':
          description: Missing x-user-id (no caller identity)
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InternalDeleteByUserRequest:
      properties:
        cognito_sub:
          type: string
          minLength: 1
          title: Cognito Sub
        user_id:
          type: string
          minLength: 1
          title: User Id
      type: object
      required:
      - cognito_sub
      - user_id
      title: InternalDeleteByUserRequest
      description: Body of `DELETE /v1/trackings/by-user`. Both identities are required.
    InternalDeleteByUserResponse:
      properties:
        deleted:
          type: integer
          title: Deleted
      type: object
      required:
      - deleted
      title: InternalDeleteByUserResponse
      description: '`200` payload of `DELETE /v1/trackings/by-user`.


        Named `deleted` rather than `count` because it says what was counted — the same

        shape Users and Orders report.'
    TrackingHistoryEntryResponse:
      properties:
        tracking_id:
          type: string
          title: Tracking Id
        user_id:
          type: string
          title: User Id
        order_id:
          type: string
          title: Order Id
        status:
          type: string
          title: Status
        datetime:
          type: string
          title: Datetime
          description: ISO-8601 UTC timestamp of this transition
      type: object
      required:
      - tracking_id
      - user_id
      - order_id
      - status
      - datetime
      title: TrackingHistoryEntryResponse
      description: 'One immutable status transition.


        Carries no `shipping_address` — the address is fixed for a tracking''s

        lifetime, so only the tracking itself would hold it (and this surface does not

        expose it at all).'
    TrackingListResponse:
      properties:
        trackings:
          items:
            $ref: '#/components/schemas/TrackingResponse'
          type: array
          title: Trackings
      type: object
      required:
      - trackings
      title: TrackingListResponse
      description: 'The batch read''s payload.


        An object with a `trackings` list rather than a bare top-level array: a bare

        array is not extensible (there is nowhere to add a field later without a

        breaking change) and is the shape most REST clients handle worst.


        There is no `total`, and no per-id error entry. Non-owned and unknown ids are

        silently omitted per the design''s ownership rule, so a count of what came back

        is exactly `len(trackings)` and anything more would start describing what the

        caller does NOT own.'
    TrackingResponse:
      properties:
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        order_id:
          type: string
          title: Order Id
        status:
          type: string
          title: Status
        datetime:
          type: string
          title: Datetime
          description: ISO-8601 UTC timestamp of the current status
        history:
          items:
            $ref: '#/components/schemas/TrackingHistoryEntryResponse'
          type: array
          title: History
      type: object
      required:
      - id
      - user_id
      - order_id
      - status
      - datetime
      - history
      title: TrackingResponse
      description: 'A tracking together with its ordered history.


        History is part of the payload rather than a separate endpoint because every

        caller of these reads wants both — the design specifies the tracking "+ its

        `Tracking_History`" for the single read and for the batch read alike.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...